clean up code, extract bdd/rspec/example_group_steps, add readme
